fix: Stf manual confirm leads threadlocal leak
feat: Introduce Stf run mode

+++ stun4j-stf-sample/stun4j-stf-boot-sample/src/test/resources/schema/mysql/ddl.sql
@@ -49,7 +49,7 @@ create table stn_stf_delay (
 --drop table if exists stn_stf_sample_mock;
 create table stn_stf_sample_mock (
   id varchar(255) not null,
-  value tinyint(3) not null,
+  value bigint(20) not null,
   primary key (id)
 );
